Add event_queue_t::poll_event_before.
---
Also adds a time member to the event_t struct for this purpose. This
uses glfwGetTime() for setting event times in callbacks, so it's
important not to call glfwSetTime after events are being polled. Doing
so will result in odd differences in event times.

If possible, never call glfwSetTime() and if you need to offset times,
keep your own base time to subtract/add to the current GLFW time.

---
sender pointer and sender_id (to identify the type of sender) also put
in event_t. In many cases, the sender will likely be ignorable.