Add _controlfp calls to set FP precision/rounding.
This is mostly to _attempt_ to ensure floating point arithmetic is the
same across platforms. This probably doesn't actually matter in single-
player games, but it should at least make it possible to later re-add
net code (currently mostly stripped when using --no-server in the build
options — would like to work on that again eventually).