client: Wrap resource loading in block to drop references first.
Avoids an error in some un-committed code where an object would release
its resources after all resources had been released. This of course is
an error because it's an attempt to release resources not held.

In addition, move the glBlendFunc/etc. calls and console system-add up
a bit closer to where context tweaks are made.