client: Move resource release to deferred function.
Avoids nesting the entire frameloop in a block and lets the function
act as a block. Since the objects will be destroyed in the correct
order (i.e., first in, last out), the deferred function should be
enough to keep this clean.

It's potentially a resource hog, but it's not in an area that matters
and I find it more readable as a result.